https://sonarcloud.io/project/issues?id=codewars&issues=AW7-YV-hI_ltG7gakBgo&open=AW7-YV-hI_ltG7gakBgo

+++ kata/7 kyu/product-of-maximums-of-array-array-series-number-2/main/Solution.java
@@ -1,7 +1,7 @@
 import static java.util.stream.IntStream.of;
 
 class Solution {
-  static long maxProduct(int[] numbers, int subSize) {
+  static long maxProduct(int[] numbers, long subSize) {
     return of(numbers).sorted().skip(numbers.length - subSize).mapToLong(i -> i).reduce(1, (p, n) -> p * n);
   }
 }
